refactor: export logic to seperate component
